claude projekte/IPO/requirements.txt
requests
plyer
python-dotenv